@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFrameworks>net472;netstandard2.0;netstandard2.1;netcoreapp2.2;netcoreapp3.1;netcoreapp5.0;netcoreapp6.0</TargetFrameworks>
+        <TargetFrameworks>net472;net6.0;net8.0;netcoreapp3.1;netstandard2.0;netstandard2.1</TargetFrameworks>
     </PropertyGroup>
 
 </Project>
